Refactored/Rearranged ticket files
